+++ src/lerobot/teleoperators/homunculus/homunculus_glove.py
@@ -180,7 +180,7 @@ class HomunculusGlove(Teleoperator):
             ValueError: any joint's recorded min and max are the same.
 
         Returns:
-            tuple[dict[str, int], dict[str, int]]: Two dictionaries *mins* and *maxes* with the extreme values
+            tuple[dict[str, float], dict[str, float]]: Two dictionaries *mins* and *maxes* with the extreme values
             observed for each joint.
         """
         if joints is None:
